refactor(rust): Move JoinOptionsIR from dsl to plans
